+++ hyperdome/common/bootstrap.py
